+++ frontend/app/src/components/dashboard/DashboardAssetTable.vue
@@ -213,7 +213,7 @@ const tableHeaders = computed<DataTableHeader[]>(() => {
       </template>
       <template #item.usdPrice="{ item }">
         <AmountDisplay
-          v-if="item.usdPrice && item.usdPrice.gte(0)"
+          :loading="!item.usdPrice || item.usdPrice.lt(0)"
           no-scramble
           show-currency="symbol"
           :price-asset="item.asset"
@@ -221,9 +221,6 @@ const tableHeaders = computed<DataTableHeader[]>(() => {
           fiat-currency="USD"
           :value="item.usdPrice"
         />
-        <div v-else class="flex justify-end">
-          <VSkeletonLoader width="70" type="text" />
-        </div>
       </template>
       <template #item.amount="{ item }">
         <AmountDisplay :value="item.amount" />